rpa/requirements.txt
# RPA Testing Tools - Dependencias ligeras (~15 MB)
# Para desarrollo y testing, no produccion

faker>=22.0.0
click>=8.1.0
pandas>=2.0.0
rich>=13.0.0
python-dateutil>=2.8.0
pytest>=7.4.0
pyyaml>=6.0.0
